refactor: remove worker pool, refactor evaluation environment in a specific module
